Refactor to use a search_controller for merchants and items - RESTful
