+++ app/controllers/api/v1/items/search_controller.rb
@@ -1,0 +1,29 @@
+class Api::V1::Items::SearchController < ApplicationController
+  def index
+    raise ActionController::ParameterMissing.new(params) if CallSearch.search_item(params)
+
+    if params[:name]
+      render json: ItemSerializer.new(Item.find_all_name(params[:name]))
+    elsif params[:min_price] && params[:max_price]
+      render json: ItemSerializer.new(Item.find_all_range(params[:min_price], params[:max_price]))
+    elsif params[:min_price]
+      render json: ItemSerializer.new(Item.find_all_min(params[:min_price]))
+    elsif params[:max_price]
+      render json: ItemSerializer.new(Item.find_all_max(params[:max_price]))
+    end
+  end
+
+  def show
+    raise ActionController::ParameterMissing.new(params) if CallSearch.search_item(params)
+
+    if params[:name]
+      render json: ItemSerializer.new(Item.find_all_name(params[:name]).first)
+    elsif params[:min_price] && params[:max_price]
+      render json: ItemSerializer.new(Item.find_all_range(params[:min_price], params[:max_price]).first)
+    elsif params[:min_price]
+      render json: ItemSerializer.new(Item.find_all_min(params[:min_price]).first)
+    elsif params[:max_price]
+      render json: ItemSerializer.new(Item.find_all_max(params[:max_price]).last)
+    end
+  end
+end
