+++ app/controllers/api/v1/items_controller.rb
@@ -32,34 +32,6 @@ class Api::V1::ItemsController < ApplicationController
     head :no_content
   end
 
-  def find
-    raise ActionController::ParameterMissing.new(params) if CallSearch.search_item(params)
-
-    if params[:name]
-      render json: ItemSerializer.new(Item.find_all_name(params[:name]).first)
-    elsif params[:min_price] && params[:max_price]
-      render json: ItemSerializer.new(Item.find_all_range(params[:min_price], params[:max_price]).first)
-    elsif params[:min_price]
-      render json: ItemSerializer.new(Item.find_all_min(params[:min_price]).first)
-    elsif params[:max_price]
-      render json: ItemSerializer.new(Item.find_all_max(params[:max_price]).last)
-    end
-  end
-  
-  def find_all
-    raise ActionController::ParameterMissing.new(params) if CallSearch.search_item(params)
-
-    if params[:name]
-      render json: ItemSerializer.new(Item.find_all_name(params[:name]))
-    elsif params[:min_price] && params[:max_price]
-      render json: ItemSerializer.new(Item.find_all_range(params[:min_price], params[:max_price]))
-    elsif params[:min_price]
-      render json: ItemSerializer.new(Item.find_all_min(params[:min_price]))
-    elsif params[:max_price]
-      render json: ItemSerializer.new(Item.find_all_max(params[:max_price]))
-    end
-  end
-
 private
   def item_params
     params.require(:item).permit(:name, :description, :unit_price, :merchant_id)
