+++ app/controllers/api/v1/merchants_controller.rb
@@ -11,25 +11,4 @@ class Api::V1::MerchantsController < ApplicationController
       render json: MerchantSerializer.new(Merchant.find(params[:id]))
     end
   end
-
-  def find
-    raise ActionController::ParameterMissing.new(params) if CallSearch.search_merchant(params)
-
-    if find_result == []
-      render json: {"data" => {}}
-    else
-      render json: MerchantSerializer.new(find_result.first)
-    end
-  end
-
-  def find_all
-    raise ActionController::ParameterMissing.new(params) if CallSearch.search_merchant(params)
-
-    render json: MerchantSerializer.new(find_result)
-  end
-
-private
-  def find_result
-    Merchant.find_all(params[:name])
-  end
 end
